Start using utils file
