+++ src/map/helpers/zone_generation/hub_generator.rs
@@ -2,32 +2,12 @@ use bevy::prelude::*;
 use bevy_ecs_tilemap::map::TilemapSize;
 use std::collections::HashMap;
 
-use super::create_tile_layout::create_hub;
+use super::{create_tile_layout::create_hub, utils::calculate_center_rect};
 use crate::map::components::{MapLayout, MapMarkers, MarkerType};
 
 const PLAYER_SPAWN_Y_OFFSET: f32 = 5.0;
 const LEVEL_EXIT_Y_OFFSET: f32 = 23.0;
 const NPC_OFFSET: f32 = 5.0;
-
-#[derive(Debug, Clone, Copy)]
-struct Rect {
-    min: Vec2,
-    max: Vec2,
-}
-
-impl Rect {
-    fn from_center_size(center: Vec2, size: Vec2) -> Self {
-        let half_size = size / 2.0;
-        Self {
-            min: center - half_size,
-            max: center + half_size,
-        }
-    }
-
-    fn center(&self) -> Vec2 {
-        (self.min + self.max) / 2.0
-    }
-}
 
 pub fn generate_hub_layout() -> MapLayout {
     let size = TilemapSize { x: 100, y: 100 };
@@ -48,11 +28,7 @@ pub fn generate_hub_layout() -> MapLayout {
 fn generate_hub_markers(map_size: TilemapSize, hub_size: TilemapSize) -> MapMarkers {
     let mut markers = HashMap::new();
 
-    // Calculate hub boundaries
-    let hub_center = Vec2::new((map_size.x / 2) as f32, (map_size.y / 2) as f32);
-
-    let hub_bounds =
-        Rect::from_center_size(hub_center, Vec2::new(hub_size.x as f32, hub_size.y as f32));
+    let hub_bounds = calculate_center_rect(map_size, hub_size);
 
     let center_of_hub = hub_bounds.center();
