+++ src/map/helpers/zone_generation/mod.rs
@@ -1,7 +1,11 @@
 mod create_tile_layout;
+mod dead_zone;
+mod hub;
 mod hub_generator;
 mod instance_generator;
 mod map_data;
+mod utils;
 mod walls;
+
 pub use hub_generator::generate_hub_layout;
 pub use instance_generator::generate_instance_layout;
